requirements.txt
kivy==2.3.0
kivymd==1.1.1
deep-translator==1.11.4
requests==2.31.0
Pillow==10.1.0
pyttsx3==2.90
gTTS==2.5.0
